feat(aula02): melhorar scripts do tema.

+++ modulo.java/aula2/src/main/java/br/com/crescer/aula2/tema/FileUtilsImpl.java
@@ -15,14 +15,14 @@ import java.util.logging.Logger;
  *
  * @author Lucas
  */
-public class FilesUtilsImpl implements FileUtils {
+public class FileUtilsImpl implements FileUtils {
 
     @Override
     public boolean mk(String string) {
         try {
             return new File(string).createNewFile();
         } catch (IOException ex) {
-            Logger.getLogger(FilesUtilsImpl.class.getName()).log(Level.SEVERE, null, ex);
+            Logger.getLogger(FileUtilsImpl.class.getName()).log(Level.SEVERE, null, ex);
             return false;
         }
     }
@@ -30,16 +30,18 @@ public class FilesUtilsImpl implements FileUtils {
     @Override
     public boolean rm(String string) {
         final File arquivo = new File(string);
-        if(arquivo.isDirectory()) {
-            System.out.println("Arquivo é inválido");
-            return false;
-        }
+        if(!arquivo.exists())
+            throw new RuntimeException("Arquivo não existe.");
+        if(arquivo.isDirectory())
+            throw new RuntimeException("Arquivo é inválido");
         return arquivo.delete();
     }
 
     @Override
     public String ls(String string) {
         final File arquivo = new File(string);
+        if(!arquivo.exists())
+            throw new RuntimeException("Arquivo não existe.");
         if(arquivo.isFile())
             return arquivo.getAbsolutePath();
         // se é diretório:
@@ -55,10 +57,8 @@ public class FilesUtilsImpl implements FileUtils {
     public boolean mv(String in, String out) {
         final File arquivoIn = new File(in);
         final File arquivoOut = new File(out);
-        if(arquivoIn.isDirectory() || arquivoOut.isDirectory()) {
-            System.out.println("Arquivo é inválido");
-            return false;
-        }
+        if(arquivoIn.isDirectory() || arquivoOut.isDirectory())
+            throw new RuntimeException("Arquivo é inválido");
         return arquivoIn.renameTo(arquivoOut);
     }
